requirements.txt
# LeafIQ - AI Plant Disease Detection System
# Core Requirements

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# AI & Machine Learning
ultralytics==8.0.196
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3

# API Integrations
google-generativeai==0.3.0
google-api-python-client==2.100.0
requests==2.31.0
wikipedia==1.4.0

# Environment & Configuration
python-dotenv==1.0.0
